+++ app/src/main/java/com/example/devproject/format/StudyInfo.kt
@@ -5,7 +5,7 @@ import android.net.Uri
 data class StudyInfo (
     var documentID : String? = null,
     var ongoing : Boolean? = null,
-    var title : String? = null,
+    var title : String = "",
     var content : String? = null,
     var offline : Boolean? = null,
     var studyURL : String? = null,
